Player Movement & Camera Follow
Player can move left and right. The character sprite turns left or right, based on its movement. The player can jump, although not so high.
The camera follows the player. It can be limited on both X and Y axis (can be used if we make it more of a side-scroller + parallax mapping)