Added Mirror Networking & Players movements + individual camera
Implemented Mirrror Networking which allows the creation of localhost servers and clients can connect to it.

I've created a build for the level for testing. You can find it in the TestBuild folder. Use only this one for now. You can open two instances of the game. Select "Host Server+Client" in one instance, and "Client" in the second one. Each player is controlled individually, together with their camera.

I tried using Cinemachine, but I couldn't sort out so the camera shows each players view - had to be done in script, but very easy.