Fixed Jittering, deleted old LoadingTestBuild
Managed to fix the jittering of the characters. It is not as smooth as it is desired to be, but it is manageable and can be worked with.

Deleted the other LoadingBuildTest and stick to a single test build folder. The updated build has the previously integrated functionalities (creating room, ready up, join level)

For Programmers:
In the Resource folder, the player character prefab, under Photon View component, there is a Interpolation field which is currently set to "Synchronize values". This is the parameter which handled and handles the jittering of the character. I tried playing around with other values (Fixed speed, lerp, etc.) but they still seem to cause the jitter. Syncing the value is the only one, so far, not causing a jitter, but an unsmooth movement. Perhaps you can play around?