Updated Player Model + Level assets organized
Updated the Player prefab with the new robot model.
Deleted old robot model.
Organized the Level Hierarchy:
All ground tileables are in Ground.
Crate and barrel tileables are in separate Crates and Barrels gameobject "folders". They are separated because, perhaps, we make them obstacles for the robot.
All the other stuff:
 DungeonBricks
 Houses, BrickTiles
 WoodPosts
 DungeonGradients
 SkyGradients
 Lamps
have been put in their old separate gameobject "folders".

Note:  dragging the robot or fairy models in the level will not actually work due to several reasons:
> There is no character distinction function yet, to distinguish the robot from the fairy, as they will need separate movement scripts
> The game will spawn the players only after going through the process of creating the room and joining it, then readying up (not do-able in editor, needs building and running 2 instances). Also, the characters that are spawned must be placed and kept in the Resources folder.

The TestBuild folder is now updated with the new level layout and player model.