Tried adding animations (failed) + fixed photonview error
I tried adding the idle and walk animations on the robot. I set the transitions correctly and it detects when the player should animate an idle state or moving state, but the character in game does not show any animation. The animator is triggered in the player movement script.
Apparently it asks for some Avatar? - check the Player prefab in Resources,  under Animator component.

Also fixed an error which occured during runtime, flooding the built game screen with errors of null exception.