Fixed Van animation & Stairs collision
Fixed Van's animation. Now the walking animation is triggered when Van is walking.
The staircase collision has been fixed, now Van can walk past the staircase without colliding with it, and Van can jump on the stairs.

Bug encountered: The elements activation/deactivation triggered by the lever is not synced for all players. Van can see the cage open, but Ava is unable to see it get out, as the collision still takes place.

To be addressed in a next update:
- Sync the Van lever pulling animation
- Sync the elements that have to be activated/deactivated

TestBuild folder is updated with the new build.