Partially Added Level Travel, not working atm
Added a trigger box which is around the ladder which spawns when the lever is pulled.
The players should travel to level 2 once both of them are inside the trigger box, however it does not happen, even though Level2 scene is in the build settings.

It might have to do with the names of the players. I will give some more tries in a bit.