Level Travelling Implemented
Level Travel now works. It is triggered when both players are at the ladder.

However, the players cannot be controlled, at least yet, in the game. I'm still looking into this. Players, however, still travel to the level and are instantiated, as the debugger shows they have joined the room.

Might have to look more into this in the near future.

For programmers;
The level travel is called within the trigger box, but the spawning of players is done in PhotonHandler script, at the bottom. In theory, it should work that way, so the connection between the players is not lost. If you, for some reason, might look into this and you may have a solution, please do try and let me know.

TestBuild server is updated with the latest build of the game.