Level Teleportation works
Finished the basis of Level Teleportation.
This will require adding two spawnpoints for each player and a level travel box per level. So if we have 3 levels, there should be 6 spawnpoints (3 for van, 3 for ava) and 3 level travel boxes.
The players are being teleported with a delay of 1 second between them so they are teleported correctly - trying to teleport both of them at the same time will only teleport Ava, and Van stays in place.
To Artists: Whenever you have time and you are ready, you can add the additional work you've done to be added in the game (backgrounds, npcs, buildings, various scenery). You can even design the level if you want, and I'll eventually place the spawnpoints and collisions.

To Miru: I'm afraid your camera limitation script has to be done again, because, when teleporting the players, the set camera boundaries impact the view the players have. Sorry.