Expanded Enemy AI behaviour types & Created Custom Test Area
Expanded Enemy AI Behaviour types:
The enemies now have 3 states. Only one state should be used for each type of enemy, depending on how they are intended to be used. As such, we have:

Move Left Right Random - when this bool is on, the enemy will randomly move left and right. It still is constrained by the Turn points. so they should be used whenever the enemies are used to randomly patrol an area.

Move Left Right Delayed - when this bool is on, the enemy detect when it hits a turn point. The enemy stops, waits for 5 seconds (the preset time, can be changed in editor), then turns the other way and moves until hits another turn point.

Follow Player - when this bool is on, the enemy disregards any turn point and constantly moves towards the robot, always the robot.

In all of these states, if the enemy touches the player, the player's position is reset. Now the fairy's position is also reset if it gets hit by the enemies.

Custom Test Area:
I've deleted the Level2 scene and created a new TestArea scene which makes testing gameplay elements & interactions easier.
This scene does not require photon, gets rid of the dialogues, and the camera always will focus the robot.
The Camera now has a boolean component devTesting which should stay on on the main camera in the Test scene. When this bool is on, the Photon View invoking is ignored.

TestBuild folder is updated with the latest version of the game.