GameManager, Level 2 layout, Level 3 added, Cogs collection
Created a GameManager which will keep track of the current level, as well as the collected cogs so far.

Overall level teleportation has been refined in order to correctly teleport the players to the designated spawnpoints. No more bugs and red text on screen and cases such as one player is teleported while the other isn't.

Level 2 layout has been completed with the platform being a platform effector (similar as the stairs in level 1), where the player can go through it from the bottom, but not from the top. This will ease things for the fairy, so the robot does not necessarily have to jump off the fairy to get on the platform.
Level box to teleport to the next level has been added.

Created a Level 3 placeholder which contains the basic elements that will be adjusted later:
Cogs - the cogs can be picked up by the player. I've set up a placeholder UI which indicated the number of cogs collected out of the total amount of cogs needed. This counter will be displayed only in the third level.

Patrolling Guard & Boulder placeholder - set up the patrolling guard standing under a platform, on top of which sits a boulder with a RigidBody component (I know it is a barrel right now, but imagine it is a boulder). Miru will work on implementing the boulder dropping on the enemy and later the enemy drops a cog.
Attached Miru's bird & cogs implementation in the placeholder level.

The TestBuild folder now contains the newest build which you can try out.

Bug found:
This one has to deal with the dialogue and timeline. Again.
When you start the game, the level 2 dialogue appears, for no reason. When the level 2 dialogue ends, the level 1 dialogue begins, but another Van appears in the scene.