Changed Lobby Menu, Added splash screens, Misc optimisations and changes.
Main changes here are the splash screens and the new lobby menu, which lists all existing rooms for the users to click on, but they can still manually type in a room name if they want to. Still no splash screen for university logo, used the darkpoint logo as placeholder for now.